+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/apis/BaseHubdleExtension.kt
@@ -1,6 +1,7 @@
 package com.javiersc.hubdle.project.extensions.apis
 
 import com.javiersc.gradle.project.extensions.withPlugins
+import com.javiersc.hubdle.project.extensions.shared.PluginId
 import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.plugins.AppliedPlugin
@@ -8,6 +9,10 @@ import org.gradle.api.plugins.AppliedPlugin
 public interface BaseHubdleExtension {
 
     public val project: Project
+
+    public fun withPlugin(pluginId: PluginId, action: Action<AppliedPlugin>) {
+        project.pluginManager.withPlugin(pluginId.id, action)
+    }
 
     public fun withPlugin(pluginId: String, action: Action<AppliedPlugin>) {
         project.pluginManager.withPlugin(pluginId, action)